bcachefs: inline bkey_cmp and in_btree_node into validate_bset_keys hot path
bch2_validate_bset_keys was making two OOL calls per key that boil
down to a handful of inline instructions each:

  - bch2_bkey_cmp_packed: the inlined version has a ~17-insn fast path
    (__bch2_bkey_cmp_packed_format_checked_inlined) when both keys
    are packed, which is the common case for prev-vs-current ordering
    checks during validation. Switch btree_node_read_bkey_cmp() to
    bch2_bkey_cmp_packed_inlined() to inline the comparison.

  - bch2_bkey_in_btree_node: two bpos comparisons plus
    bkey_fsck_err_on() error handling. Move from bkey_methods.c to
    btree/read.h as a static inline; both callers (bset_key_validate
    and interior.c's split-point check) already include read.h.

Production-build hot path now does the comparison and in-node check
inline with no OOL call overhead. Debug builds (CONFIG_BCACHEFS_DEBUG=y)
gain two __bkey_unpack_pos calls per key from the EBUG_ON inside
__bch2_bkey_cmp_packed_format_checked_inlined, but that's just
inlining the per-key debug check that the OOL cmp_packed previously
hid; the production build avoids it entirely.
